+++ examples/bickley_jet_cmc_cg.jl
@@ -11,6 +11,7 @@ import ClimaCore: Fields, Domains, Topologies, Meshes, Spaces
 import ClimaCore.Operators
 import ClimaCore.Geometry
 
+#=
 # set up boilerplate
 include("../src/interface/WIP_domains.jl")
 include("../src/interface/WIP_models.jl")
@@ -24,6 +25,25 @@ include("../src/backends/climacore/tendencies.jl")
 include("../src/interface/WIP_boundary_conditions.jl")
 include("../src/interface/WIP_simulations.jl")
 include("../src/interface/WIP_physics.jl")
+=#
+using ClimaAtmos
+using ClimaAtmos.Utils
+using ClimaAtmos.Interface
+using ClimaAtmos.Backends
+
+# explicit imports from Interface
+import ClimaAtmos.Interface: PeriodicRectangle
+import ClimaAtmos.Interface: BarotropicFluidModel
+import ClimaAtmos.Interface: TimeStepper
+import ClimaAtmos.Interface: DirichletBC, Simulation
+
+# explicit imports from Backends
+import ClimaAtmos.Backends: create_ode_problem, evolve
+
+# External Stuff
+using IntervalSets
+using OrdinaryDiffEq: ODEProblem, solve, SSPRK33
+
 
 # set up parameters
 parameters = (
