+++ src/interface/WIP_simulations.jl
@@ -8,38 +8,8 @@ struct Simulation{A,B,C,D,E} <: AbstractSimulation
     ode_problem::E
 end
 
-function Simulation(
-    backend::AbstractBackend;
-    model,
-    timestepper::AbstractTimestepper,
-    callbacks,
-)
-    ode_problem = create_ode_problem(
-        backend, 
-        model,
-        timestepper, 
-    )
-  
-    return Simulation(
-        backend,
-        model,
-        timestepper, 
-        callbacks, 
-        ode_problem,
-    )
-end
 
-function evolve(simulation::Simulation{<:ClimaCoreBackend})
-    return solve(
-        simulation.ode_problem,
-        simulation.timestepper.method,
-        dt = simulation.timestepper.dt,
-        saveat = simulation.timestepper.saveat,
-        progress = simulation.timestepper.progress, 
-        progress_message = simulation.timestepper.progress_message,
-    )
-end
-
+#=
 function evolve(simulation::Simulation{<:DiscontinuousGalerkinBackend})
     method        = simulation.timestepper.method
     start         = simulation.timestepper.tspan[1]
@@ -71,3 +41,4 @@ function evolve(simulation::Simulation{<:DiscontinuousGalerkinBackend})
         )
     end
 end
+=#
